refactor(user-management): relocate api_key.py to autobot_shared, cosmetic diff only (#12647)
